% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iiplot.R
\name{iiplot}
\alias{iiplot}
\title{Influence  interest plot}
\usage{
iiplot(
  dat,
  sub.id = c(1, 6),
  title = "",
  min.influ = 1,
  max.influ = 5,
  min.inter = 1,
  max.inter = 5,
  standard = FALSE,
  idcol = NULL,
  idminim = NULL,
  idmaxim = NULL
)
}
\arguments{
\item{dat}{An input of data set (\emph{see} \strong{Details}).}

\item{sub.id}{A column id of each dimension (\emph{see} \strong{Details}).}

\item{title}{A title for the plot (\emph{see} \strong{Details}).}

\item{min.influ}{A minimum score of influence (\emph{see} \strong{Details}).}

\item{max.influ}{A maximum score of influence (\emph{see} \strong{Details}).}

\item{min.inter}{A minimum score of interest (\emph{see} \strong{Details}).}

\item{max.inter}{A maximum score of interest (\emph{see} \strong{Details}).}

\item{standard}{A logical value for standard/ original (\emph{see} \strong{Details}).}

\item{idcol}{A numeric (vectors) for id column (\emph{see} \strong{Details}).}

\item{idminim}{A numeric (vectors) for minimum score (\emph{see} \strong{Details}).}

\item{idmaxim}{A numeric (vectors) for maximum score (\emph{see} \strong{Details}).}
}
\value{
Function returns a list of coordinate scores and
two dimension plot.
}
\description{
This function plot the influence-interest
dimensions for governance leve (stakeholder analysis).
}
\details{
The data set is a n x c data frame or matrix object. The influence-interest
plot draws only two dimensions of influence (x) and interest (y).
For \code{sub.id}, user has to input the first column of each dimension,
for example \code{sub.id=c(1,6)} meaning that the influence indicators are
column 1 to 5 while the rests are the interest indicators.
Minimum and maximum scores for each dimension are also defined in the
\code{min.influ}, \code{max.influ}, \code{min.inter}, and \code{max.inter}.
}
\examples{
#data simulation
data("sldata")
iiplot(sldata)

}
\references{
Bryson, J. M. 2004.
What to do when Stakeholders matter.
Public Management Review, 6(1), 21–53.
doi:https://doi.org/10.1080/14719030410001675722
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
