% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IiScale15n10}
\alias{IiScale15n10}
\title{Influence interest simulation data set 1}
\format{
A data frame with 10 rows and 10 variables:
\describe{
  \item{manag}{management authority.}
  \item{budget}{budget contribution.}
  \item{inst}{institutional capacity.}
  \item{depen}{dependency level.}
  \item{colab}{collaboration intensity.}
  \item{invol}{program involvement.}
  \item{benef}{perceived benefits.}
  \item{author}{management responsibility.}
  \item{parti}{stakeholder participationn.}
  \item{impact}{program impacts.}

}
}
\source{
simulated data.
}
\usage{
IiScale15n10
}
\description{
A simulation dataset contains ten indicators.
It has 10 variables of 10 locations measured for their scores/ indices.
The minimum score is 1, and the maximum is 5.
}
\keyword{datasets}
