% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leverage.R
\name{leverage}
\alias{leverage}
\title{Leverage plot}
\usage{
leverage(datori, datres, title = "", minim = 0, maxim = 10, zoom.in = TRUE)
}
\arguments{
\item{datori}{An input of data set (\emph{see} \strong{Details}).}

\item{datres}{An input of rapfish result (\emph{see} \strong{Details}).}

\item{title}{A title for the plot.}

\item{minim}{A minimum score (\emph{see} \strong{Details}).}

\item{maxim}{A maximum score (\emph{see} \strong{Details}).}

\item{zoom.in}{A logical input for a zoom in leverage plot.}
}
\value{
Function returns list of leverage scores and
a plot of leverage.
}
\description{
This function plot the leverage of Rapfish
}
\details{
\code{datori} is the original data set, while
\code{datres} is the corresponding rapfist result from the rapfish function.
For \code{minim} and  \code{maxim} arguments, they have default
values 0 and 10, respectively. A user can modify them.
}
\examples{
#data simulation
data("social")
res <- rapfish(social)
leverage(social,res,title = "social")

}
\references{
Kavanagh, P., & Pitcher, T. J. 2004. Implementing Microsoft Excel software for Rapfish:
a technique for the rapid appraisal of fisheries status.
doi:http://dx.doi.org/10.14288/1.0074801
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
