% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lvi.R
\name{lvi}
\alias{lvi}
\title{Livelihood vulnerability index}
\usage{
lvi(
  datlvi,
  sub.id = c(1, 3, 5, 7, 9),
  dim.name = c("human", "natur", "physic", "financial", "social")
)
}
\arguments{
\item{datlvi}{An input of data set (\emph{see} \strong{Details}).}

\item{sub.id}{An column id for each component (\emph{see} \strong{Details}).}

\item{dim.name}{Components' name (\emph{see} \strong{Details}).}
}
\value{
Function returns a list of livelihood vulnerability indices.
}
\description{
This function calculates Livelihood vulnerability index
}
\details{
The data set is a data frame. The columns in \code{datlvi}
indicate variables where the rows are the districts/ objects.
\code{sub.id} indicates the column id of the first variable for
each sub-component.
}
\examples{
#data simulation
data("social")
lvi(social,sub.id = c(1),dim.name = c("Social"))
lvi(three, sub.id = c(1,9,16),dim.name = c("Social","Institution","Technology"))

}
\references{
Micah B. Hahn, Anne M. Riederer, Stanley O. Foster. 2009.
The Livelihood Vulnerability Index: A pragmatic approach to assessing
risks from climate variability and change—A case study in Mozambique.
Global Environmental Change, Volume 19, Issue 1, Pages 74-88.
doi:http://doi.org/10.1016/j.gloenvcha.2008.11.002
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
