% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LviEqualScale15n10}
\alias{LviEqualScale15n10}
\title{Lvi simulation data set 1}
\format{
A data frame with 10 rows and 30 variables:
\describe{
  \item{human1}{Indicator 1 of human dimension.}
  \item{human2}{Indicator 2 of human dimension.}
  \item{natur1}{Indicator 1 of natur dimension.}
  \item{natur2}{Indicator 2 of natur dimension.}
  \item{phys1}{Indicator 1 of physic dimension.}
  \item{phys2}{Indicator 2 of physic dimension.}
  \item{fin1}{Indicator 1 of financial dimension.}
  \item{fin2}{Indicator 2 of financial dimension.}
  \item{soc1}{Indicator 1 of social dimension.}
  \item{soc2}{Indicator 2 of social dimension.}

}
}
\source{
simulated data.
}
\usage{
LviEqualScale15n10
}
\description{
A simulation dataset contains five dimensions, i.e. human, natur,
physic, financial, an social dimensions.
It has 10 variables of 10 locations measured for their scores/ indices.
For each dimension consists two indicators (equal number of indicators).
The minimum score is 1, and the maximum is 5.
}
\keyword{datasets}
