% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnormal.R
\name{mc_normal}
\alias{mc_normal}
\title{Montecarlo normal method}
\usage{
mc_normal(n = 10, dat, minim = 0, maxim = 10)
}
\arguments{
\item{n}{A number of data simulation (\emph{see} \strong{Details}).}

\item{dat}{An input of data set (\emph{see} \strong{Details}).}

\item{minim}{A minimum score (\emph{see} \strong{Details}).}

\item{maxim}{A maximum score (\emph{see} \strong{Details}).}
}
\value{
Function returns coordinates of Montecarlo simulation.
}
\description{
This function generate montecarlo simulation with normal method.
}
\details{
Because the normal method applying box-mueller transformation
\code{n} will be always converted to even number even it the input is an
odd number.
The data set \code{dat} arguments is a data set as a source
data for the simulation.
For \code{maxim} and  \code{minim} arguments, they have default
values 10 and 0, respectively. A user can modify them.
}
\examples{
#data simulation
data("social")
b2 <- mc_normal(10,social)
b2

}
\references{
Kavanagh, P., & Pitcher, T. J. 2004. Implementing Microsoft Excel software for Rapfish:
a technique for the rapid appraisal of fisheries status.
doi:http://dx.doi.org/10.14288/1.0074801
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
