% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RapEqualScale15n1000}
\alias{RapEqualScale15n1000}
\title{Rapfish simulation data set 3}
\format{
A data frame with 100 rows and 15 variables:
\describe{
  \item{ecol1}{Indicator 1 of ecological dimension.}
  \item{ecol2}{Indicator 2 of ecological dimension.}
  \item{econ1}{Indicator 1 of economic dimension.}
  \item{econ2}{Indicator 2 of economic dimension.}
  \item{soc1}{Indicator 1 of social dimension.}
  \item{soc2}{Indicator 2 of social dimension.}
  \item{tech1}{Indicator 1 of technology dimension.}
  \item{tech2}{Indicator 2 of technology dimension.}
  \item{inst1}{Indicator 1 of institution dimension.}
  \item{inst2}{Indicator 2 of institution dimension.}

}
}
\source{
simulated data.
}
\usage{
RapEqualScale15n1000
}
\description{
A simulation dataset contains five dimensions, i.e. ecological, economic,
social, technology, and institution dimensions.
It has 10 variables of 1000 locations measured for their scores/ indices.
Each dimension consists two indicators (equal number of indicators).
The minimum score is 1, and the maximum is 5.
}
\keyword{datasets}
