% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfish.R
\name{rapfish}
\alias{rapfish}
\title{Rapfish calculation}
\usage{
rapfish(datrap, num_fish = 12, minim = 0, maxim = 10)
}
\arguments{
\item{datrap}{An input of data set (\emph{see} \strong{Details}).}

\item{num_fish}{A number of fish (\emph{see} \strong{Details}).}

\item{minim}{A minimum score (\emph{see} \strong{Details}).}

\item{maxim}{A maximum score (\emph{see} \strong{Details}).}
}
\value{
Function returns a list with \emph{7} length indicates the
fish/ object coordinates, stress value, variance explained per component,
good-bad-up-down coordinate, anchors coordinates, maximum, and minimum values.
}
\description{
This function calculates Rapfish coordinates.
}
\details{
The data set is a data frame. The columns indicate variables where
the rows are the fish/ objects. \code{num_fish} indicates the number of
fish/ objects. For \code{minim} and  \code{maxim} arguments, they have default
values 0 and 10, respectively. A user can modify them.
}
\examples{
#data simulation
data("social")
a <- rapfish(social)
a

}
\references{
Kavanagh, P., & Pitcher, T. J. 2004. Implementing Microsoft Excel software for Rapfish:
a technique for the rapid appraisal of fisheries status.
doi:http://dx.doi.org/10.14288/1.0074801
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
