% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapfish2.R
\name{rapfish2}
\alias{rapfish2}
\title{Multidimensional Rapfish calculation}
\usage{
rapfish2(
  datrap,
  num_fish = 10,
  minim = 0,
  maxim = 10,
  iddim = c(1, 6, 10),
  standard = FALSE,
  idcol = NULL,
  idminim = NULL,
  idmaxim = NULL,
  dim.name = NULL
)
}
\arguments{
\item{datrap}{An input of data set (\emph{see} \strong{Details}).}

\item{num_fish}{A number of fish (\emph{see} \strong{Details}).}

\item{minim}{A minimum score (\emph{see} \strong{Details}).}

\item{maxim}{A maximum score (\emph{see} \strong{Details}).}

\item{iddim}{A A numeric (vectors) for id column of each dimension (\emph{see} \strong{Details}).}

\item{standard}{A logical value for standard/ original (\emph{see} \strong{Details}).}

\item{idcol}{A numeric (vectors) for id column (\emph{see} \strong{Details}).}

\item{idminim}{A numeric (vectors) for minimum score (\emph{see} \strong{Details}).}

\item{idmaxim}{A numeric (vectors) for maximum score (\emph{see} \strong{Details}).}

\item{dim.name}{A character (vectors) for each dimension name (\emph{see} \strong{Details})..}
}
\value{
Function returns a list with \emph{7} length indicates the
fish/ object coordinates, stress value, variance explained per component,
good-bad-up-down coordinate, anchors coordinates, maximum, and minimum values.
}
\description{
This function calculates Rapfish coordinates.
}
\details{
The data set is a data frame. The columns indicate variables where
the rows are the fish/ objects. \code{num_fish} indicates the number of
fish/ objects. For \code{minim} and  \code{maxim} arguments, they have default
values 0 and 10, respectively. A user can modify them.

\code{standard} is a logical value with TRUE or FALSE where TRUE can be
applied for heterogenous scales among variables. \code{idcol} indicates
the column number of the heterogeneous scales. When heterogenous scales
are applied, \code{idminim} and \code{idmaxim} indicate the minimum and
maximum scores for the heterogenous scale, respectively. \code{idcol},
\code{idminim} and \code{idmaxim} must have an identical length.
}
\examples{
#data simulation
data("RapEqualScale15n10")
rest <- rapfish2(RapEqualScale15n10,num_fish = 10,minim=1,maxim=5,iddim = c(1,7,13,19,25))
rest

}
\references{
Kavanagh, P., & Pitcher, T. J. 2004. Implementing Microsoft Excel software for Rapfish:
a technique for the rapid appraisal of fisheries status.
doi:http://dx.doi.org/10.14288/1.0074801
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
