% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapplot.R
\name{rapplot}
\alias{rapplot}
\title{Rapfish plot}
\usage{
rapplot(datplot, title = "", minim = 0, maxim = 10)
}
\arguments{
\item{datplot}{An input of rapfish result (\emph{see} \strong{Details}).}

\item{title}{A title for the plot.}

\item{minim}{A minimum score (\emph{see} \strong{Details}).}

\item{maxim}{A maximum score (\emph{see} \strong{Details}).}
}
\value{
Function returns a plot of rapfish result.
}
\description{
This function plot the Rapfish result
}
\details{
The data set is a rapfist result from rapfish function.
For \code{minim} and \code{maxim} arguments, they have default
values 0 and 10, respectively. A user can modify them.
}
\examples{
#data simulation
data("social")
a <- rapfish(social)
rapplot(a, title = "social")

}
\references{
Kavanagh, P., & Pitcher, T. J. 2004. Implementing Microsoft Excel software for Rapfish:
a technique for the rapid appraisal of fisheries status.
doi:http://dx.doi.org/10.14288/1.0074801
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
