% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapsummary.R
\name{rapsummary}
\alias{rapsummary}
\title{Rapfish summary for montecarlo simulation}
\usage{
rapsummary(datbox, title = "", scatter = TRUE)
}
\arguments{
\item{datbox}{An input of montecarlo result (\emph{see} \strong{Details}).}

\item{title}{A title for the plot.}

\item{scatter}{A logical input for a scatter plot or a circled-boxplot.}
}
\value{
Function returns a plot of montecarlo summary.
}
\description{
This function plot the montcarlo plot summary
}
\details{
The data set is a rapfist result from rapfish function.
For \code{maxim} and  \code{minim} arguments, they have default
values 10 and 0, respectively. A user can modify them.
}
\examples{
#data simulation
data("social")
b1 <- mc_triangular(10,social)
b2 <- mc_normal(10,social)
rapsummary(b1)
rapsummary(b2)

}
\references{
Kavanagh, P., & Pitcher, T. J. 2004. Implementing Microsoft Excel software for Rapfish:
a technique for the rapid appraisal of fisheries status.
doi:http://dx.doi.org/10.14288/1.0074801
}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
