% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RapUnequalScale15n10}
\alias{RapUnequalScale15n10}
\title{Rapfish simulation data set 7}
\format{
A data frame with 10 rows and 26 variables:
\describe{
  \item{ecol1}{Indicator 1 of ecological dimension.}
  \item{ecol2}{Indicator 2 of ecological dimension.}
  \item{ecol3}{Indicator 3 of ecological dimension.}
  \item{ecol4}{Indicator 4 of ecological dimension.}
  \item{ecol5}{Indicator 5 of ecological dimension.}
  \item{econ1}{Indicator 1 of economic dimension.}
  \item{econ2}{Indicator 2 of economic dimension.}
  \item{econ3}{Indicator 3 of economic dimension.}
  \item{econ4}{Indicator 4 of economic dimension.}
  \item{econ5}{Indicator 5 of economic dimension.}
  \item{econ6}{Indicator 6 of economic dimension.}
  \item{econ7}{Indicator 7 of economic dimension.}
  \item{econ8}{Indicator 8 of economic dimension.}
  \item{soc1}{Indicator 1 of social dimension.}
  \item{soc2}{Indicator 2 of social dimension.}
  \item{soc3}{Indicator 3 of social dimension.}
  \item{soc4}{Indicator 4 of social dimension.}
  \item{soc5}{Indicator 5 of social dimension.}
  \item{soc6}{Indicator 6 of social dimension.}
  \item{tech1}{Indicator 1 of technology dimension.}
  \item{tech2}{Indicator 2 of technology dimension.}
  \item{tech3}{Indicator 3 of technology dimension.}
  \item{tech4}{Indicator 4 of technology dimension.}
  \item{inst1}{Indicator 1 of institution dimension.}
  \item{inst2}{Indicator 2 of institution dimension.}
  \item{inst3}{Indicator 3 of institution dimension.}
}
}
\source{
simulated data.
}
\usage{
RapUnequalScale15n10
}
\description{
A simulation dataset contains five dimensions, i.e. ecological, economic,
social, technology, and institution dimensions.
It has 26 variables of 10 locations measured for their scores/ indices.
Each dimension consists an unequal number of indicators.
The minimum score is 1, and the maximum is 5.
}
\keyword{datasets}
