% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RapUnequalScale15n1000}
\alias{RapUnequalScale15n1000}
\title{Rapfish simulation data set 9}
\format{
A data frame with 1000 rows and 29 variables:
\describe{
  \item{ecol1}{Indicator 1 of ecological dimension.}
  \item{ecol2}{Indicator 2 of ecological dimension.}
  \item{ecol3}{Indicator 3 of ecological dimension.}
  \item{ecol4}{Indicator 4 of ecological dimension.}
  \item{ecol5}{Indicator 5 of ecological dimension.}
  \item{ecol6}{Indicator 6 of ecological dimension.}
  \item{ecol7}{Indicator 7 of ecological dimension.}
  \item{econ1}{Indicator 1 of economic dimension.}
  \item{econ2}{Indicator 2 of economic dimension.}
  \item{econ3}{Indicator 3 of economic dimension.}
  \item{econ4}{Indicator 4 of economic dimension.}
  \item{econ5}{Indicator 5 of economic dimension.}
  \item{soc1}{Indicator 1 of social dimension.}
  \item{soc2}{Indicator 2 of social dimension.}
  \item{soc3}{Indicator 3 of social dimension.}
  \item{soc4}{Indicator 4 of social dimension.}
  \item{soc5}{Indicator 5 of social dimension.}
  \item{soc6}{Indicator 6 of social dimension.}
  \item{soc7}{Indicator 7 of social dimension.}
  \item{soc8}{Indicator 8 of social dimension.}
  \item{tech1}{Indicator 1 of technology dimension.}
  \item{tech2}{Indicator 2 of technology dimension.}
  \item{tech3}{Indicator 3 of technology dimension.}
  \item{tech4}{Indicator 4 of technology dimension.}
  \item{tech5}{Indicator 5 of technology dimension.}
  \item{tech6}{Indicator 6 of technology dimension.}
  \item{inst1}{Indicator 1 of institution dimension.}
  \item{inst2}{Indicator 2 of institution dimension.}
  \item{inst3}{Indicator 3 of institution dimension.}
}
}
\source{
simulated data.
}
\usage{
RapUnequalScale15n1000
}
\description{
A simulation dataset contains five dimensions, i.e. ecological, economic,
social, technology, and institution dimensions.
It has 29 variables of 1000 locations measured for their scores/ indices.
Each dimension consists an unequal number of indicators.
The minimum score is 1, and the maximum is 5.
}
\keyword{datasets}
