% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sldata}
\alias{sldata}
\title{Influence–interest stakeholder data set}
\format{
A data frame with 12 rows and 8 variables:
\describe{
  \item{Influ.1}{Influence indicator 1: influence in the management.}
  \item{Influ.2}{Influence indicator 2: budget contribution.}
  \item{Influ.3}{Influence indicator 3: institution capacity}
  \item{Influ.4}{Influence indicator 4: dependentcy level.}
  \item{Influ.5}{Influence indicator 5: cooperation intensity.}
  \item{Inter.1}{Interest indicator 1: engagement.}
  \item{Inter.2}{Interest indicator 2: existing benefit.}
  \item{Inter.3}{Interest indicator 3: autority.}
  \item{Inter.4}{Interest indicator 4: program.}
  \item{Inter.5}{Interest indicator 5: program impact.}
}
}
\source{
simulated data.
}
\usage{
sldata
}
\description{
A simulation dataset containing two component, i.e. influence and
interest dimensions. It has five objects of stakeholders and
10 variables consisting of both 5 influence and interest indicators.
The minimum score is 0, and the maximum is 5.
}
\keyword{datasets}
