% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{social}
\alias{social}
\title{Social dimension data set}
\format{
A data frame with 12 rows and 8 variables:
\describe{
  \item{educ}{Education index.}
  \item{prosper}{Prosperity potential.}
  \item{care}{Social caring.}
  \item{suitable}{Index of suitability.}
  \item{otherAct}{Other activities.}
  \item{Wpoktan}{Group of farming.}
  \item{Wmanage}{Group management.}
  \item{conflict}{Potential of conflict.}
}
}
\source{
simulated data.
}
\usage{
social
}
\description{
A simulation dataset containing an component, i.e. social dimension.
It has eight variables of 12 locations measured for their scores/ indices.
The minimum score is 0, and the maximum is 10.
}
\keyword{datasets}
