% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{three}
\alias{three}
\title{Social, institution, and technology dimensions data set}
\format{
A data frame with 12 rows and 20 variables:
\describe{
  \item{educ}{Education index.}
  \item{prosper}{Prosperity potential.}
  \item{care}{Social caring.}
  \item{suitable}{Index of suitability.}
  \item{otherAct}{Other activities.}
  \item{Wpoktan}{Group of farming.}
  \item{Wmanage}{Group management.}
  \item{conflict}{Potential of conflict.}
  \item{participate}{Participation index.}
  \item{intensity}{Instensity of activities.}
  \item{subtitute}{Subtitute organization.}
  \item{cooperation}{Index of cooperation.}
  \item{market}{Potential market.}
  \item{ikm}{small enterprise.}
  \item{integrate}{Integration a new information.}
  \item{maintain}{Maintainance activities.}
  \item{acceptance}{Acceptance index.}
  \item{management}{Management index.}
  \item{poultry}{Poultry activites.}
  \item{fertilizer}{Fertilizer usage.}
}
}
\source{
simulated data.
}
\usage{
three
}
\description{
A simulation dataset containing three components, i.e. social,
institution, and technology dimensions.
It has eight variables/ indicators of 12 locations measured for
their scores/ indices. Social indicators are column number 1 to 8,
institution variables are column number 9 to 15, and
the rest are technology indicators.
The minimum score is 0, and the maximum is 10.
}
\keyword{datasets}
