% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarplot.R
\name{radarplot}
\alias{radarplot}
\title{Radar plot of sustainability livelihood analysis}
\usage{
radarplot(dat, title = "", lvi.out = TRUE, object = NULL)
}
\arguments{
\item{dat}{An input of data set (\emph{see} \strong{Details}).}

\item{title}{A title for the plot (\emph{see} \strong{Details}).}

\item{lvi.out}{An indicator whether the input data from lvi function (\emph{see} \strong{Details}).}

\item{object}{A numeric input for the object id (\emph{see} \strong{Details}).}
}
\value{
Function returns a radar plot indicating the vulnerability score.
}
\description{
This function plot the sustainability livelihood in radar
plot depending on the included number of dimensions.
}
\details{
The data set is a n x 2 data frame or matrix object. The sustainability
livelihood radar plot draws all included dimensions.
For \code{title}, user can input additional information for the plot title.
For \code{lvi.out} is indicator input whether \code{TRUE} or
\code{FALSE} while \code{object} must be a numeric with a length of 1.
It is valid only \code{lvi.out = TRUE}.
}
\examples{
#data simulation
data("three")
rest <- lvi(three, sub.id = c(1,9,16),dim.name = c("Social","Institution","Technology"))
radarplot(rest,object = 5)

#non lvi output
datex <- data.frame(
dimension = c("Social", "Institution","Technology"),
score = c(5,7,9))
radarplot(datex,lvi.out = FALSE)

}
\author{
Weksi Budiaji \cr Contact: \email{budiaji@untirta.ac.id}
}
